code: Fixed BadCode BUG
Failed validation of arguments was falling through to the "DO IT" code
block.  SYMPTOM: Invalid numerators or denomiators were also registering
as divide by zero errors.

Fixed.

SIDE NOTE: The TediousStart framework highlighted this bug.  If BadCode
was supposed to test the test code, does this mean that BadCode
successfully verified TediousStart or that TediousStart successfully
tested BadCode?!  Either way, it's a success.

Changes to be committed:
	modified:   badcode/main.py