TEST-11: Initial commit TFT.run_test() func tests
Initial commit of the TediousFuncTest.run_test() functional tests.
This is a work in progress.  I haven't even designed the test cases yet.
However, I did realize I'd need to capture the output and then test it.
I've done that before so I decided to put a pin in this and work on
that.

SPOILER ALERT: The test_redirect_std_streams work is about 2/3 done.

Changes to be committed:
	new file:   test/functional_tests/test_tediousfunctest_run_test.py

Untracked files:
	test/unit_tests/test_redirect_std_streams.py