TEST-11: Redirect test case initial commit
NOTE: I may be using unit_tests but this feels more like a functional
test.

The test class is working as expected.  28 test cases written and
working.  All tests passing but normal 2 (and the related test cases)
highlighted an odd edge case when using whitespace characters in
conjunction with file-based streams.  The output even looks weird when
printed with DEBUGGING statements.  If I didn't know better, I'd think
there were some backspace characters in that output.

For the time being, I backed off the salting for the file-based stream
tests and explicitly added Special 16 to help get to the bottom of
what's happening.

All the Normal, Error, and Boundary test cases are implemented.
The Special test cases have been designed but need to be implemented.
Then, after RedirectStdStreams() is in an acceptable state, go back to
working on the TediousFuncTest.run_test() test cases.

EOD Commit

Changes to be committed:
	new file:   test/unit_tests/test_redirect_std_streams.py