TEST-11: RSStreams() testing complete 43/44 pass
43 of 44 RedirectStdStreams tests pass.
The last one, Special 16, is being skipped.
Special 16 highlights an edge-case BUG when using a file-based stream to
read the \r character.  The output is not as expected.  This may be a
result of running this test on a platform that doesn't natively use \r.
Regardless, it represents an extreme edge case.  A warning was added to
the class documentation.

TO DO: DON'T DO NOW... add a backlog ticket to investigate further.

Changes to be committed:
	modified:   test/unit_tests/test_redirect_std_streams.py

Changes not staged for commit:
	modified:   tediousstart/redirect_std_streams.py