TEST-23: Refactored shell script env var responses
Shell scripts now test for the existence of the environment variable
before exporting it.  This alleviates some issues with nested script
execution.  It also allows the user to control the verbosity level
without having to modify the shell script.

Also, added usage instructions to the file comment blocks.

Changes to be committed:
	modified:   devops/scripts/test_installation.sh
	modified:   devops/scripts/test_local.sh
	modified:   devops/scripts/test_local_examples.sh